Add support for injectively reading []
Adding support for injectively reading the empty list.  This is a
feature request by @renzo-orsini in madnificent/jsown#10

We are not enabling this feature in the injective reader yet, as it may
be breaking for other users.  First doublechecking if we should have
this feature enabled in the injective reader or not.