Abide order of specified keys for programatically created objects
When a json object is parsed and written, the order of the keys is
correct.  However, when you add keys to an object, the keys are added to
the front of the object, rather than to the back.  This is indeed
somewhat faster, but it doesn't feel nice when looking at the resulting
json files.  We're switching this around for nicer json output.