Allow reading of integers to contain non-regular whitespace
The issue of reading with a return character in the string was
mentioned in #26.

This change idea allows for setting the :junk-allowed option when
reading numbers.  This option allows for more freedom then necessary,
but the performance implications seem very minimal.  Although we
should consider alternatives, setting the (unexported)
`jsown::*allow-junk-in-integers-p*` parameter to a truethy value will
allow such reading at the moment.  As we ponder on alternatives a
better solution may be found.

It is currently optional as there might be side-effects and there is
probably a cleaner option.  The current solution allows for random
junk where we should probably only allow whitespace.  Nevertheless,
the solution may help should someone be stuck.